Remove GO_REPLACE.txt and references to Go vendoring
The OpenFaaS golang-http-template has had changes made due to Go modules and vendoring not
being able to be used at the same time. To prevent errors during the build process stating that the go.mod
and vendor/modules.txt are out of sync only GO111MODULE is required. During the build a script is used
that will automatically handle replacements.